+++ src/js/common/component.js
