--- src/js/letters.js
@@ -1,1 +1,0 @@
-export default 'abcdefghijklmnopqrstuvwxyzабвгдеёжзийклмнопрстуфхцчшщыъьэюя';
